+++ pillarbox-demo/src/main/java/ch/srgssr/pillarbox/demo/ui/showcases/layouts/SimpleStory.kt
@@ -24,7 +24,7 @@ import androidx.media3.common.Player
 import ch.srgssr.pillarbox.demo.shared.data.DemoItem
 import ch.srgssr.pillarbox.demo.shared.data.Playlist
 import ch.srgssr.pillarbox.demo.shared.di.PlayerModule
-import ch.srgssr.pillarbox.player.PillarboxPlayer
+import ch.srgssr.pillarbox.player.exoplayer.PillarboxExoPlayer
 import ch.srgssr.pillarbox.ui.ScaleMode
 import ch.srgssr.pillarbox.ui.widget.player.PlayerSurface
 
@@ -57,7 +57,7 @@ fun SimpleStory() {
 
 /**
  * Simple story player
- * Each DemoItem have a [PillarboxPlayer], the player is released onDispose
+ * Each DemoItem have a [PillarboxExoPlayer], the player is released onDispose
  *
  * @param demoItem The DemoItem to play
  * @param isPlaying to pause or play the player
