+++ pillarbox-player/src/androidTest/java/ch/srgssr/pillarbox/player/IsPlayingAllTypeOfContentTest.kt
@@ -10,6 +10,7 @@ import androidx.media3.common.PlaybackException
 import androidx.media3.common.Player
 import androidx.media3.common.util.ConditionVariable
 import androidx.test.platform.app.InstrumentationRegistry.getInstrumentation
+import ch.srgssr.pillarbox.player.exoplayer.PillarboxExoPlayer
 import ch.srgssr.pillarbox.player.utils.ContentUrls
 import org.junit.Assert
 import org.junit.Test
@@ -31,10 +32,10 @@ class IsPlayingAllTypeOfContentTest {
     fun isPlayingTest() {
         // Context of the app under test.
         val appContext = getInstrumentation().targetContext
-        val atomicPlayer = AtomicReference<PillarboxPlayer>()
+        val atomicPlayer = AtomicReference<PillarboxExoPlayer>()
         val waitIsPlaying = WaitIsPlaying()
         getInstrumentation().runOnMainSync {
-            val player = PillarboxPlayer(
+            val player = PillarboxExoPlayer(
                 appContext
             )
             atomicPlayer.set(player)
