+++ pillarbox-player/src/main/java/ch/srgssr/pillarbox/player/PillarboxExoPlayer.kt
@@ -4,35 +4,17 @@
  */
 package ch.srgssr.pillarbox.player
 
-import androidx.media3.common.Player
 import androidx.media3.exoplayer.ExoPlayer
-import androidx.media3.exoplayer.SeekParameters
 
 /**
  * Pillarbox [ExoPlayer] interface extension.
  */
-interface PillarboxExoPlayer : ExoPlayer {
-
+interface PillarboxExoPlayer : PillarboxPlayer, ExoPlayer {
     /**
-     * Listener
+     * Handle audio focus with currently set AudioAttributes
+     * @param handleAudioFocus true if the player should handle audio focus, false otherwise.
      */
-    interface Listener : Player.Listener {
-        /**
-         * On smooth seeking enabled changed
-         *
-         * @param smoothSeekingEnabled The new value of [smoothSeekingEnabled]
-         */
-        fun onSmoothSeekingEnabledChanged(smoothSeekingEnabled: Boolean)
+    fun setHandleAudioFocus(handleAudioFocus: Boolean) {
+        setAudioAttributes(audioAttributes, handleAudioFocus)
     }
-
-    /**
-     * Smooth seeking enabled
-     *
-     * When [smoothSeekingEnabled] is true, next seek events is send only after the current is done.
-     *
-     * To have the best result it is important to
-     * 1) Pause the player while seeking.
-     * 2) Set the [ExoPlayer.setSeekParameters] to [SeekParameters.CLOSEST_SYNC].
-     */
-    var smoothSeekingEnabled: Boolean
 }
