+++ pillarbox-player/src/main/java/ch/srgssr/pillarbox/player/service/PillarboxMediaSessionService.kt
@@ -11,13 +11,14 @@ import androidx.media3.common.Player
 import androidx.media3.session.MediaSession
 import androidx.media3.session.MediaSessionService
 import ch.srgssr.pillarbox.player.PillarboxPlayer
+import ch.srgssr.pillarbox.player.exoplayer.PillarboxExoPlayer
 import ch.srgssr.pillarbox.player.utils.PendingIntentUtils
 
 /**
  * `PillarboxMediaSessionService` implementation of [MediaSessionService].
  * It is the recommended way to make background playback for Android.
  *
- * It handles only one [MediaSession] with one [PillarboxPlayer].
+ * It handles only one [MediaSession] with one [PillarboxExoPlayer].
  *
  * Usage:
  * Add these permissions inside your manifest:
@@ -64,8 +65,8 @@ abstract class PillarboxMediaSessionService : MediaSessionService() {
      * @param mediaSessionCallback The MediaSession.Callback to use [MediaSession.Builder.setCallback].
      */
     fun setPlayer(
-        player: PillarboxPlayer,
         mediaSessionCallback: MediaSession.Callback = object : DefaultMediaSessionCallback {}
+        player: PillarboxExoPlayer,
     ) {
         if (this.player == null) {
             this.player = player
