+++ pillarbox-player/src/test/java/ch/srgssr/pillarbox/player/PillarboxExoPlayerMediaItemTest.kt
@@ -11,6 +11,7 @@ import androidx.media3.exoplayer.DefaultLoadControl
 import androidx.media3.test.utils.FakeClock
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import ch.srgssr.pillarbox.player.exoplayer.PillarboxExoPlayer
 import ch.srgssr.pillarbox.player.extension.getCurrentMediaItems
 import org.junit.Before
 import org.junit.runner.RunWith
@@ -18,13 +19,13 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 
 @RunWith(AndroidJUnit4::class)
-class PillarboxPlayerMediaItemTest {
-    private lateinit var player: PillarboxPlayer
+class PillarboxExoPlayerMediaItemTest {
+    private lateinit var player: PillarboxExoPlayer
 
     @Before
     fun createPlayer() {
         val context = ApplicationProvider.getApplicationContext<Context>()
-        player = PillarboxPlayer(
+        player = PillarboxExoPlayer(
             context = context,
             seekIncrement = SeekIncrement(),
             loadControl = DefaultLoadControl(),
